backend/requirements.txt
fastapi
uvicorn
pyjwt
httpx
PyYAML
passlib[bcrypt]
psycopg[binary]
